+++ game.c
@@ -1,0 +1,19 @@
+#include <pic32mx.h>
+#include <stdint.h>
+#include "project.h"
+
+
+void gameLoop(){
+    while(1){
+        display_dots();
+    }
+
+}
+
+void createCharacter(){
+    
+}
+
+
+
+
