+++ project.h
@@ -1,0 +1,8 @@
+
+
+void display_init();
+
+void display_clear();
+
+void display_dots();
+
